home view model and home view refactor
