+++ FitnessApp/Home/Models/Activity.swift
@@ -1,0 +1,22 @@
+//
+//  Activity.swift
+//  FitnessApp
+//
+//  Created by poonam mittal on 28/07/24.
+//
+
+import Foundation
+import SwiftUI
+
+
+//creating model for the activity
+
+struct Activity {
+    let id: Int
+    let title: String
+    let subtitle: String
+    let image: String
+    let tintColor: Color
+    let amount: String
+}
+
