+++ FitnessApp/Home/Models/Workout.swift
@@ -1,0 +1,19 @@
+//
+//  Workout.swift
+//  FitnessApp
+//
+//  Created by poonam mittal on 28/07/24.
+//
+
+import SwiftUI
+
+struct Workout {
+    let id: Int
+    let title: String
+    let image: String
+    let tint: Color
+    let duration: String
+    let date: String
+    let calories: String
+}
+
