+++ FitnessApp/Home/Views/ActivityCard.swift
@@ -7,18 +7,6 @@
 
 import SwiftUI
 
-//creating model for the activity
-
-struct Activity {
-    let id: Int
-    let title: String
-    let subtitle: String
-    let image: String
-    let tintColor: Color
-    let amount: String
-}
-
-
 
 struct ActivityCard: View {
