+++ FitnessApp/Home/Views/HomeView.swift
@@ -7,27 +7,12 @@
 
 import SwiftUI
 
+
 struct HomeView: View {
     
-    @State var calories: Int = 123
-    @State var active: Int = 56
-    @State var stand: Int = 8
+    @StateObject var viewModel = HomeViewModel()
     
-    var mockActivities = [
-        Activity(id: 0, title: "Today Steps", subtitle: "Goal 12,000", image: "figure.walk", tintColor: .green, amount: "9,810"),
-        Activity(id: 1, title: "Today", subtitle: "Goal 1,000", image: "figure.walk", tintColor: .red, amount: "812"),
-        Activity(id: 2, title: "Today Steps", subtitle: "Goal 12,000", image: "figure.walk", tintColor: .blue, amount: "9,812"),
-        Activity(id: 3, title: "Today Steps", subtitle: "Goal 50,000", image: "figure.run", tintColor: .purple, amount: "55,812")
-    
-    ]
-    
-    var mockWorkout = [
-        Workout(id: 0, title: "Running", image: "figure.run", tint: .cyan, duration: "51 mins", date: "Aug 1" , calories: "512 kcal"),
-        Workout(id: 1, title: "Strength Training", image: "figure.run", tint: .red, duration: "51 mins", date: "Aug 3"  , calories: "341 kcal"),
-        Workout(id: 2, title: "Running", image: "figure.run", tint: .purple, duration: "51 mins", date: "Aug 11"  , calories: "341 kcal"),
-        Workout(id: 3, title: "Running", image: "figure.run", tint: .cyan, duration: "51 mins", date: "Aug 13"  , calories: "341 kcal")
-    ]
-    
+   
     
     var body: some View {
         NavigationStack {
@@ -40,14 +25,14 @@ struct HomeView: View {
                     HStack {
                         
                         Spacer()
-                        VStack {
+                        VStack(alignment: .leading) {
                             VStack(alignment: .leading, spacing: 8) {
                                 Text("Calories")
                                     .font(.callout)
                                     .bold()
                                     .foregroundColor(.red)
                                 
-                                Text("123 kcal")
+                                Text("\(viewModel.calories)")
                                     .bold()
                             }
                             .padding(.bottom)
@@ -59,7 +44,7 @@ struct HomeView: View {
                                     .bold()
                                     .foregroundColor(.green)
                                 
-                                Text("52 mins")
+                                Text("\(viewModel.active)")
                                     .bold()
                             }
                             .padding(.bottom)
@@ -70,17 +55,17 @@ struct HomeView: View {
                                     .bold()
                                     .foregroundColor(.blue)
                                 
-                                Text("8 hours")
+                                Text("\(viewModel.stand)")
                                     .bold()
                             }
                         }
                         Spacer()
                         
                         ZStack {
-                            ProgressCircleView(progress: $calories, goal: 600, color: .red)
-                            ProgressCircleView(progress: $active, goal: 60, color: .green)
+                            ProgressCircleView(progress: $viewModel.calories, goal: 600, color: .red)
+                            ProgressCircleView(progress: $viewModel.active, goal: 60, color: .green)
                                 .padding(.all, 20)
-                            ProgressCircleView(progress: $stand, goal: 12, color: .blue)
+                            ProgressCircleView(progress: $viewModel.stand, goal: 12, color: .blue)
                                 .padding(.all, 40)
                         }
                         .padding(.horizontal)
@@ -106,7 +91,7 @@ struct HomeView: View {
                     .padding(.horizontal)
                     
                     LazyVGrid(columns: Array(repeating: GridItem(spacing: 20), count: 2)) {
-                        ForEach(mockActivities, id: \.id) { activity in
+                        ForEach(viewModel.mockActivities, id: \.id) { activity in
                             ActivityCard(activity: activity)
                             
                         }
@@ -133,7 +118,7 @@ struct HomeView: View {
                     .padding(.top)
                 
                     LazyVStack() {
-                        ForEach(mockWorkout, id: \.id) { workout in
+                        ForEach(viewModel.mockWorkout, id: \.id) { workout in
                             WorkoutCard(workout: workout)
                             
                         }
