+++ FitnessApp/Home/Views/WorkoutCard.swift
@@ -7,16 +7,6 @@
 
 import SwiftUI
 
-struct Workout {
-    let id: Int
-    let title: String
-    let image: String
-    let tint: Color
-    let duration: String
-    let date: String
-    let calories: String
-}
-
 
 
 struct WorkoutCard: View {
